feat: add demographics to stats page
